Restructured the java rest api to be more domain driven and be split into three parts: - api (general api classes) - registration - prediction
Fixed error handling, documentation and more
Made methods not be dishonest and promote immutability
