+++ server/src/main/java/META-INF/MANIFEST.MF
@@ -1,3 +1,3 @@
 Manifest-Version: 1.0
-Main-Class: com.dynabyte.marleyjavarestapi.MarleyJavaRestApiApplication
+Main-Class: com.dynabyte.marleyrest.MarleyJavaRestApiApplication
