--- server/src/main/java/com/dynabyte/marleyjavarestapi/MarleyJavaRestApiApplication.java
@@ -1,16 +1,0 @@
-package com.dynabyte.marleyjavarestapi;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-/**
- * Runs the rest api spring boot application
- */
-@SpringBootApplication
-public class MarleyJavaRestApiApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(MarleyJavaRestApiApplication.class, args);
-    }
-
-}
