--- server/src/main/java/com/dynabyte/marleyjavarestapi/facerecognition/exception/PersonAlreadyInDbException.java
@@ -1,12 +1,0 @@
-package com.dynabyte.marleyjavarestapi.facerecognition.exception;
-
-/**
- * Exception for when trying to register a person who is already in the person database
- */
-public class PersonAlreadyInDbException extends RuntimeException {
-
-    public PersonAlreadyInDbException(String message) {
-        super(message);
-    }
-
-}
