--- server/src/main/java/com/dynabyte/marleyjavarestapi/facerecognition/model/FaceRecognitionResult.java
@@ -1,17 +1,0 @@
-package com.dynabyte.marleyjavarestapi.facerecognition.model;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.Data;
-
-@Data
-public class FaceRecognitionResult {
-
-    @JsonProperty
-    private boolean isFace;
-    private String faceId;
-
-    public FaceRecognitionResult(boolean isFace, String faceId) {
-        this.isFace = isFace;
-        this.faceId = faceId;
-    }
-}
