+++ server/src/main/java/com/dynabyte/marleyrest/MarleyJavaRestApiApplication.java
@@ -1,0 +1,16 @@
+package com.dynabyte.marleyrest;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+/**
+ * Runs the rest api spring boot application
+ */
+@SpringBootApplication
+public class MarleyJavaRestApiApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(MarleyJavaRestApiApplication.class, args);
+    }
+
+}
