+++ server/src/main/java/com/dynabyte/marleyrest/api/exception/ResponseBodyNotFoundException.java
@@ -1,0 +1,11 @@
+package com.dynabyte.marleyrest.api.exception;
+
+/**
+ * Exception for when a response body is not supplied in a request
+ */
+public class ResponseBodyNotFoundException extends RuntimeException {
+
+    public ResponseBodyNotFoundException(String message) {
+        super(message);
+    }
+}
