+++ server/src/main/java/com/dynabyte/marleyrest/prediction/request/ImageRequest.java
@@ -1,0 +1,18 @@
+package com.dynabyte.marleyrest.prediction.request;
+
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+/**
+ * Request including an image in base64 format
+ */
+@Data
+@NoArgsConstructor
+public class ImageRequest {
+
+    private String image;
+
+    public ImageRequest(String image) {
+        this.image = image;
+    }
+}
