+++ server/src/main/java/com/dynabyte/marleyrest/registration/exception/MissingPersonInDbException.java
@@ -1,0 +1,12 @@
+package com.dynabyte.marleyrest.registration.exception;
+
+/**
+ * Exception for when a faceId is found during a prediction but a corresponding person is missing in the database
+ */
+public class MissingPersonInDbException extends RuntimeException {
+
+    public MissingPersonInDbException(String message) {
+        super(message);
+    }
+
+}
