+++ server/src/main/java/com/dynabyte/marleyrest/registration/request/LabelPutRequest.java
@@ -1,0 +1,20 @@
+package com.dynabyte.marleyrest.registration.request;
+
+import lombok.Data;
+
+/**
+ * Request to python face recognition api
+ * Includes an image in base64 format and a faceId to add the encoded image to as a reference
+ */
+@Data
+public class LabelPutRequest {
+
+    String image;
+    String faceId;
+
+    public LabelPutRequest(String image, String faceId) {
+        this.image = image;
+        this.faceId = faceId;
+    }
+
+}
