--- server/src/test/java/com/dynabyte/marleyjavarestapi/MarleyJavaRestApiApplicationTests.java
@@ -1,13 +1,0 @@
-package com.dynabyte.marleyjavarestapi;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class MarleyJavaRestApiApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
