+++ server/src/test/java/com/dynabyte/marleyrest/MarleyJavaRestApiApplicationTests.java
@@ -1,0 +1,13 @@
+package com.dynabyte.marleyrest;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class MarleyJavaRestApiApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
